Add missing changes
